requirements.txt
Appium-Python-Client
selenium
pytest
Ruff